+++ kitty/window.py
@@ -1524,6 +1524,13 @@ class Window:
         text = process_remote_print(msg)
         print(text, end='', flush=True)
 
+    def handle_restore_cursor_appearance(self, msg: memoryview | None = None) -> None:
+        opts = get_options()
+        self.screen.cursor.blink = opts.cursor_blink_interval[0] != 0
+        self.screen.cursor.shape = opts.cursor_shape
+        self.screen.cursor_visible = True
+        delattr(self.screen.color_profile, 'cursor_color')
+
     def send_cmd_response(self, response: Any) -> None:
         self.screen.send_escape_code_to_child(ESC_DCS, '@kitty-cmd' + json.dumps(response))